HWY-170: Move `Endorsement` types around.
