+++ node/src/components/consensus/highway_core/endorsement.rs
@@ -1,4 +1,3 @@
-use serde::{Deserialize, Serialize};
 use thiserror::Error;
 
 use crate::components::consensus::traits::Context;
@@ -16,11 +15,7 @@ pub(crate) enum EndorsementError {
 
 /// Testimony that creator of `vote` was seen honest
 /// by `endorser` at the moment of creating this endorsement.
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(bound(
-    serialize = "C::Hash: Serialize",
-    deserialize = "C::Hash: Deserialize<'de>",
-))]
+#[derive(Clone, Debug, Eq, PartialEq)]
 pub(crate) struct Endorsement<C: Context> {
     /// Vote being endorsed.
     vote: C::Hash,
@@ -45,11 +40,7 @@ impl<C: Context> Endorsement<C> {
 
 /// Testimony that creator of `vote` was seen honest
 /// by `endorser` at the moment of creating this endorsement.
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(bound(
-    serialize = "C::Hash: Serialize",
-    deserialize = "C::Hash: Deserialize<'de>",
-))]
+#[derive(Clone, Debug, Eq, PartialEq)]
 pub(crate) struct SignedEndorsement<C: Context> {
     /// Original endorsement,
     endorsement: Endorsement<C>,
@@ -72,38 +63,8 @@ impl<C: Context> SignedEndorsement<C> {
     pub(crate) fn validator_idx(&self) -> ValidatorIndex {
         self.endorsement.creator
     }
-}
 
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(bound(
-    serialize = "C::Hash: Serialize",
-    deserialize = "C::Hash: Deserialize<'de>",
-))]
-pub(crate) struct Endorsements<C: Context> {
-    pub(crate) vote: C::Hash,
-    pub(crate) endorsers: Vec<(ValidatorIndex, C::Signature)>,
-}
-
-impl<C: Context> Endorsements<C> {
-    pub fn new<I: IntoIterator<Item = SignedEndorsement<C>>>(endorsements: I) -> Self {
-        let mut iter = endorsements.into_iter().peekable();
-        let vote = *iter.peek().expect("non-empty iter").vote();
-        let endorsers = iter
-            .map(|e| {
-                assert_eq!(e.vote(), &vote, "endorsements for different votes.");
-                (e.validator_idx(), e.signature)
-            })
-            .collect();
-        Endorsements { vote, endorsers }
-    }
-
-    /// Returns hash of the endorsed vode.
-    pub fn vote(&self) -> &C::Hash {
-        &self.vote
-    }
-
-    /// Returns an iterator over validator indexes that endorsed the `vote`.
-    pub fn validator_ids(&self) -> impl Iterator<Item = &ValidatorIndex> {
-        self.endorsers.iter().map(|(v, _)| v)
+    pub(crate) fn signature(&self) -> &C::Signature {
+        &self.signature
     }
 }
