+++ node/src/components/consensus/highway_core/highway/vertex.rs
@@ -5,7 +5,7 @@ use serde::{Deserialize, Serialize};
 use crate::{
     components::consensus::{
         highway_core::{
-            endorsement::Endorsements,
+            endorsement::SignedEndorsement,
             evidence::Evidence,
             state::{self, Panorama},
             validators::ValidatorIndex,
@@ -166,5 +166,39 @@ impl<C: Context> WireVote<C> {
     /// Returns the time at which the round containing this vote began.
     pub(crate) fn round_id(&self) -> Timestamp {
         state::round_id(self.timestamp, self.round_exp)
+    }
+}
+
+#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
+#[serde(bound(
+    serialize = "C::Hash: Serialize",
+    deserialize = "C::Hash: Deserialize<'de>",
+))]
+pub(crate) struct Endorsements<C: Context> {
+    pub(crate) vote: C::Hash,
+    pub(crate) endorsers: Vec<(ValidatorIndex, C::Signature)>,
+}
+
+impl<C: Context> Endorsements<C> {
+    pub fn new<I: IntoIterator<Item = SignedEndorsement<C>>>(endorsements: I) -> Self {
+        let mut iter = endorsements.into_iter().peekable();
+        let vote = *iter.peek().expect("non-empty iter").vote();
+        let endorsers = iter
+            .map(|e| {
+                assert_eq!(e.vote(), &vote, "endorsements for different votes.");
+                (e.validator_idx(), e.signature().clone())
+            })
+            .collect();
+        Endorsements { vote, endorsers }
+    }
+
+    /// Returns hash of the endorsed vode.
+    pub fn vote(&self) -> &C::Hash {
+        &self.vote
+    }
+
+    /// Returns an iterator over validator indexes that endorsed the `vote`.
+    pub fn validator_ids(&self) -> impl Iterator<Item = &ValidatorIndex> {
+        self.endorsers.iter().map(|(v, _)| v)
     }
 }
